design: approve request timeout design
Final changes

- Support for idle timeout has been dropped. We can revisit this when
there is specific customer demand.
- Retry count is defined to be a number, the problems of the zero value
for NumRetries coming from a failed string parsing, or actually being
the number 0 are resolved by specifing that 0 is the "not set" value for
NumRetries. This means you cannot explicitly say
```
   timeout:
     count: 0
```
But it turns out that leaving the field out has the same effect.
- Retry on status code support is hard coded to "50x". We can revisit
this when there is specific customer demand.

Signed-off-by: Dave Cheney <dave@cheney.net>