remove AES128-* and AES256-* from permitted ciphers list
The full list of ciphers that envoy permits by default are listed at [1].

The defaults rate an "A" on qualy's ssl labs, however the AES128-* and AES256-*
ciphers are marked as "weak". This appears to be for a combination of two reasons:

1. They don't support "Forward Secrecy" [2]
2. They are easy to implement wrongly, exposing users to the ROBOT[3] attack.
   Presumably boringssl (which envoy uses) has implemented them correctly, but
   it seems the general recommendation is to avoid them if possible, just in case [4]

Removing these ciphers doesn't improve the overall qualy's score, but it does
remove the "weak" flags from the list of supported ciphers.

In practice, this is unlikely to hurt client compatibility for users of
contour. Most of the clients that might need to use the removed ciphers (like
windows XP, java 6+7, and android <= 4.3) are unable to connect without TLS
1.0, which contour never enables.

This cipher list is also ignored when a connection negotiates TLS 1.3.
I'm sure why that is, but the envoy 1.10 docs [5] seems pretty clear:

> this setting has no effect when negotiating TLS 1.3

Signed-off-by: James Healy <james@yob.id.au>

[1] https://www.envoyproxy.io/docs/envoy/v1.9.1/api-v2/api/v2/auth/cert.proto.html?highlight=ciphers
[2] https://en.wikipedia.org/wiki/Forward_secrecy
[3] https://robotattack.org/
[4] https://community.qualys.com/thread/17971-tlsrsawithaes256cbcsha-comes-to-be-weak-cipher#comment-39790
[5] https://www.envoyproxy.io/docs/envoy/v1.10.0/api-v2/api/v2/auth/cert.proto.html?highlight=ciphers