internal/grpc: increment version_info on each response
Updates #499
Updates #273

Previously each response sent to Envoy set the version_info field to
"0". The xDS documentation says that the version_info field should
change if the result returned by the management server has changed.
The xDS protocol does not place constraints on what the version_info
field should contain, other than it is a string. Internally the
Cache.Register method maintains a count of updates to the cache, so this
PR uses the string form of this counter for the version_info field in
the response.

The nonce field on the response message is used in the ACK/NACK protocol
from Envoy back to the management server which Contour doesn't
implement, but it seems reasonable to also use the string form of the
cache counter in both the version_info and nonce fields.

Signed-off-by: Dave Cheney <dave@cheney.net>