internal/dag: dag.Routes now dispatch to a set of dag.Clusters
Insert a dag.Cluster between the dag.Route and the dag.HTTPService it
dispatches too. This creates a place for Cluster specific configuration
that is neither global to the Route, nor a property of the backend k8s
service. The Cluster holds the properties that apply to traffic leaving
the Route towards the K8s service.

Also, thread dag.Cluster into the dot debug visualisation.

Signed-off-by: Dave Cheney <dave@cheney.net>