Add prometheus annotations for contour container
The deployments has annotations to scrape metrics only for envoy. This
patch adds annotations to scrape metrics for the contour container too.

The annotations are added to the contour k8s service. This is for two
reasons:

 * The pod spec annotations only support to specify one container,
AFAIK, and are already used for metrics for the envoy container

 * This seems like the easiest way to work-around the just mentioned
annotations limitation while being a trivial patch and that works out of
the box on all setups.

With this patch, metrics for the contour container even when envoy and
contour are co-located in the same pod work out of the box.

Updates: #1035

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>