internal/dag: Missing retry count implies 1 retry
Updates #1019

The retry policy documentation (#1042) states that a `retryPolicy:`
stanza with no fields implies a retry count of 1 -- if you don't want to
retry, don't add a `retryPolicy:` key.

This PR addresses this issue, and breaks out the retryPolicy logic into
a new file with associated tests to isolate this buisiness logic.

Signed-off-by: Dave Cheney <dave@cheney.net>