internal/dag: unify timeout and retry policy logic
Updates #1019

Rename timeoutPolicyIngressRoute and friend to timeoutPolicy as they
represent the current best practice logic for handling timeout/retries.

The older annotation based logic is handled differently because if the
differences in not found vs no set logic with respect to retry-on and
timeout parsing.

There are some TODO's that highlight the inconsistencies in the logic,
its not clear if its worth taking a compatability break to change that
logic now. There is good code coverage in this area that ensures we
won't regress the logic so maybe just leaving well enough alone is good
enough.

Signed-off-by: Dave Cheney <dave@cheney.net>