internal/envoy: roll back request timeout
I'm on the fence about the request timeout limit, I think 15 seconds is
reasonable, but if there are people with large uploads then this will
blow them up. I really don't want to make this value configurable
because its complicated to make it configurable per vhost, and I also
don't want to expose it to ingressroute users.

We'll revisit this in a separate issue.

Signed-off-by: Dave Cheney <dave@cheney.net>