internal/dag: respect force-ssl-redirect over allow-http annotation
Fixes #1023

The behaviour when these two annotations are both specified is somewhat
surprising: `allow-http: false` meant that no routes were not registered
for port 80, even if `force-ssl-redirect: true` was set - leading to a
404 where a 3xx with an upgrade to `https` was expected.

With this commit, we prioritize respecting `force-ssl-redirect`: if this
annotation is specified and set to `true`, we always register a port 80
route for the ingress, even if `allow-http: false`, so that the forced
upgrade can take effect.

Signed-off-by: Cera Davies <ceralena.davies@gmail.com>

squash! internal/dag: respect force-ssl-redirect over allow-http annotation

Document the new annotation precedence.