Update metrics examples for Envoy v1.10
Since the recent upgrade to envoy 1.10 there is no need to use the
statsd sink, as envoy exports those metrics by default to prometheus
now.

This was added on envoy PR:
https://github.com/envoyproxy/envoy/pull/5601 that is included in 1.10.

This commits just removes the statsd references and updates the doc.

With these changes, the statsd-enable contour flag can probably be
removed on a future patch, will create an issue for that.

Fixes: #1035
Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>