internal/envoy: move accesslog filter into its own file
Upates #876

Move accesslog fitler generation into its own file and function.
This moves the gross accesslog renamed imports to their own file to hide
that horribleness.

Also this removes the accesslog helper which returned a
types.StructValue as those functions will be rewritten to return
TypedConfigs in the next PR.

Signed-off-by: Dave Cheney <dave@cheney.net>