internal/envoy: convert TCPProxy to TypedConfig
Updates #876
Updates #499

Convert TCPProxy listener filter to a TypedConfig. This gives us
protection if Envoy changes the proto definition of the filter types in
the future. It also reduces the background churn of idempotent, but not
bit-for-bit identical messages between Contour and Envoy.

This has lots of pleasent follow on effects; sorting of TCP Proxy
members is less gross, and we remove all uses of the odeous
messageToStruct via JSON hack.

Also, add tests for FileAccessLog and HTTPConnectionManager. They're
simple, but by adding a test we can then be sure that we can use them in
other functions as we have a test to assert what they are returning --
and if we change it in the future.

Signed-off-by: Dave Cheney <dave@cheney.net>