Makefile: promote unparam to check target
Promote mvdan's unparam to run as part of the make check target. This
would have spotted a condition in a recent PR where a paremeter was
declared in the signture but wasn't used in the body of the function.

Skip the api/ directory as the generated files aren't lint clean -- but
we don't really care because a computer wrote them for us.

Signed-off-by: Dave Cheney <dave@cheney.net>