internal/grpc: move TypeURL into internal/contour
Updates #1091

To convert a proto.Message to a proto.Any you need to know its declared
type (not Go type, protobuf type). Previously this was handled by the
CDS/EDS/RDS/etc types in internal/grpc, but this isn't very useful as
those types are tightly coupled to the values that are sourced from
internal/contour. Thus splitting the type that declares TypeURL and the
type that provides proto.Messages of those types permits things to be
constructed inconsistently. With this change the thing that provides
proto.Message values, via the Values() method, also provides a TypeURL
method to describe the type of the things provided by Values.

Also, add missing StreamSecret test case.

Signed-off-by: Dave Cheney <dave@cheney.net>