internal/contour: add unit tests for ListenerCache.{Contents,Query}
Similar to SecretsCache there is no valid blank valid for an unknown
listener. This is because each listener has a required address field.
Making up a value there could cause Envoy to open that port, or worse,
conflict with an already open port. Because there is no safe valid
value, when querying for an unknown listener, return nothing.

Signed-off-by: Dave Cheney <dave@cheney.net>