all: remove Maglev and RingHash as supported load balancer strategies
Fixes #1150
Fixes #1030
Updates #1005

RingHash and Maglev are two balancing / affinity strategies that can be
configured per Cluster, however they are not useful by themselves.
Without the ability to configure the hash key, which is usually some for
of session af session or cookie, these strategies are not useful and
cannot be used correctly.

As such, remove them from the list of selectable options. Support for
session affinity will be added later.

Signed-off-by: Dave Cheney <dave@cheney.net>