examples: simplify validation for spec.virtualhost.fqdn
Fixes #1117

We've tried a few times to find a regex that fits the domain names that
people want to use but there are always edge cases.

Give up and simply require that the fqdn field is required. Ultimately
if the string isn't a valid DNS name, nothing will be able to route to
it, so the cost of writing nonsense here is low.

Signed-off-by: Dave Cheney <dave@cheney.net>