cmd/contour: refactor bootstrap flag and command handling
This PR is the start of a series attempting to counter the bloat of
cmd/contour's main.main function. It established a pattern where we pass
into a helper function the kingpin Application object. The helper
registers the flags it needs and returns a context type object which
holds the references to all the flags needed by the subcommand. The
context object is later passed back to the function executing the
subcommand so it can pick up the context it established with its flags.
This is a little verbose, but preferable to pooping global variables
everywhere.

Additionally, the config.Namespace envvar helper was replaced with a
flag on bootstrap called `--namespace` which takes its value from Envar
first, then the flag secondly.

Signed-off-by: Dave Cheney <dave@cheney.net>