cmd/contour: move setupTLSContext to a method on serveCtx
As all of the parameters for setupTLSContext were supplied by
serveContext, move the helper there. The replacement method, tlsconfig
is safe to call multiple times -- while the values from calling this
method twice won't be identical, they come from the same source, so are
effectively equal (not that we'd do that anyway, but it's smart to plan
ahead).

While here, clean up the error handling in tlsconfig -- and actually
check the error from loading a cert.

Lastly simplify the "is tls enabled check" to be just "is the value from
tlsconfig nil, if so, no tls"

Signed-off-by: Dave Cheney <dave@cheney.net>