cmd/contour: add -c config-path flag
Fixes #1130

This PR adds the ability to supply configuration values to contour serve
from a configuration file. Some gymnastics are required to make this
work with kingpin and keep the precidence of config file -> env vars ->
cli flags.

At the moment only a few values are plumbed into the config file, mainly
to keep the linter's happy. More will be added in 0.15.

Signed-off-by: Dave Cheney <dave@cheney.net>