internal/dag: move IngressClass filtering to dag.KubernetesCache
Updates #1166
Updates #1172

Move the filtering of ingress class from the k8s resource event handler
to the dag KuberenetesCache. The cache now returns a boolean indicating
if the insert or remove operation caused the cache to change. This in
turn sets the ground rules for whether a dag update will take place in
the caller.

Signed-off-by: Dave Cheney <dave@cheney.net>