internal/contour: move dag.KubernetesCache logging to the caller
Updates #1166
Updates #1172

This PR moves the lock that prevents the cache from being updated during
dag rebuild from the dag package to the caller. This is extremely poor
form, the lock of a thing shouldn't be mutated externally, but in this
case its a stepping stone to a better implementation that doesn't need a
lock.

Signed-off-by: Dave Cheney <dave@cheney.net>