internal/contour: merge ResourceEventHandler and HoldoffNotifier
Updates #1166
Updates #1172

Merge the holdoff notifier into contour's resource event handler. This
ensures that we're always using the holdoff notifier, although by
default the holdoff delays default to zero unless configured. This
causes the regeneration to fire immediately during tests while
cmd/contour provides the standard 100ms and 500ms holdoff delays.

This PR is a mid point between a rewrite of the handler/holdoff notifier
to own the kubernetes cache which will remove the need to lock it, and
allow us to use the cache when deciding if notification should occur.

Signed-off-by: Dave Cheney <dave@cheney.net>