cmd/contour: cleanly shutdown grpc server
Fixes #1361

The stop channel was not wired into the grpc server stop signal. This
meant unless the grpc server was the instigator contour serve would hang
on group shutdown.

This PR fixes this by wiring stop into grpc.Server.Stop. It also adjusts
the stopping message in startInformer to be stopped so it is the same as
the other workers. This is important because we don't issue a "stopped"
message thus the reader could be confused that what was stopping had not
yet stopped.

Signed-off-by: Dave Cheney <dave@cheney.net>