internal/dag: simplify Builder's virtualhost map
Fixes #1135

This PR removes the Builder's listeners map by storing two maps of
virtualhost and securevirtualhost objects directly in the builder. This
resolves #1135 by hard coding that there will only ever be a listener
for port 80 and a listener for port 443 -- this might not be the case
forever, but it will be for 1.0 and we can revisit that later.

In the mean time this significantly simplifies the handling of listeners
inside the builder. There is no more messy map storing two kinds of
objects which must be carefully identified every time we walk through
the map. It simplifies the dag.Listener object which no longer needs to
store things in a map when a slice will do. And it simplifies dag.dag()
because we no longer delete from a map while iterating through it, we
simply copy the value virtualhost objects into a fresh listener as part
of building the dag.Dag value.

Signed-off-by: Dave Cheney <dave@cheney.net>