internal/e2e: fix invalid fixture
Updates #1159

The names of ports in a service is defined by the service, _not_ the pod
spec. Thus it is not possible for a namespace+name+portname to refer to
more than one port number. The foo/9999 foo/8000 example cannot happen
in the wild because it is not possible to construct Service document
that has two named ports, Name: foo, Port 9999 AND Name: foo, Port:
8000.

As this condition is not possible, replace it by moving the port to the
NotReadySet and adjust the expected output.

Signed-off-by: Dave Cheney <dave@cheney.net>