internal/contour: sort endpoints with a ClusterLoadAssignment
Fixes ##1159

Simplify the conversion of Endpoint document to ClusterLoadAssignment
set. This simplification makes it easier to sort the list of endpoints
without a ClusterLoadAssignment. ClusterLoadAssignments are sorted
themselves on the way out of EndpointTranslator.Query and Contents.

The simplification takes advantage of the fact that a service namespace,
name, port and port name tupple only appers once per EndpointSet. The
same port and port name (if there is more than one port) cannot appear
on the service document more than once therefore for each endpoint
subset, we'll never see the same port mentioned more than once.

Signed-off-by: Dave Cheney <dave@cheney.net>