internal/dag: add HTTPLoadbalancer to KubernetesCache
Fixes #1411

Add support for projectcontour.HTTPLoadBalancer and
projectcontour.TLSCertificateDelegation to the KubernetesCache. They
should now be available for building a DAG.

There is one commented out test case to do with
projectcontour.HTTPLoadBalancer Upstream certificate validation as the
route matching syntax is not yet defined.

ingressroute.TLSCertificateDelegations and
projectcontour.TLSCertificateDelegations are stored seperately but
merged together when computing the set of secrets to rebuild on. This
effectively makes them the same from Contour's point of view.
ingressroute's TLSCertificateDelegation can be used to delegate a secret
to an projectcontour.HTTPLoadBalancer and vice versa. Given that
ingressroute will be deprecated by (possibly removed) by Contour 1.0
this is probably a reasonable compromise.

Signed-off-by: Dave Cheney <dave@cheney.net>