internal/dag: reject delegations from any ingressroute to a root
Fixes #865

Since 0.6 we've had a method called `dag.Builder.validIngressRoutes`
which excluded all but the IngressRoute roots. However we didn't have a
check at delegation point that the delegate was not itself a root. This
PR fixes that and adds some belt and braces tests to assert it won't
happen.

Adjust status tests to deal with the various order of delegation checks
vs other status checks.

Signed-off-by: Dave Cheney <dave@cheney.net>