Dockerfile: upgrade to Go 1.13
Fixes #1377

BREAKING CHANGE

We build with -mod=readonly so that the build will fail if anything is
out of date in our go.mod file. golang/go#30667 changed the behaviour of
the go tool to require that go.mod AND go.sum be up to date so this
means I cannot use my hack of .gitignoring the go.sum file.

This means when you pull this commit you will probably have to delete
your local go.sum working copy. If we're lucky then you shouldn't need
to change it after that. If we're not lucky, the bloody thing will
always be dirty and we'll waste a lot of time telling people to revert
their changes to this file. C'est la vie.

Signed-off-by: Dave Cheney <dave@cheney.net>