internal/dag: permit combining non tls routing with tcp proxy
Fixes #910

The use case is for a service offering tls passthrough on port 443, the
same service may wish to offer a non tls port which informs users they
need to connect over port 443. Presumably port 443 is not HTTPS but some
other protocol, the port 80 fallback is a convenience for people typing
the address into their browsers.

Signed-off-by: Dave Cheney <dave@cheney.net>