go.mod: upgrade to envoyproxy/go-control-plane v0.9.0
Fixes #1488

In the process move as much of the proto type boilerplace to its own
package, internal/protobuf.

Similarly resolve a lot of the casting from int to uint to int32. This
exposed a case where the types that envoy would accept would not accept
signed integers, so correct the generated API to rejected negative
weights and ports. This means we don't need to write status checks for
them :)

Signed-off-by: Dave Cheney <dave@cheney.net>