internal/dag: simplify parseAnnotation
parseAnnotation returns uint32 values, its input was previously a map
and the key in the map to parse. This is unnecessary as the caller has
to supply both the map and the map key just to pass a string to
parseAnnotaion.

Simplify this to parseUInt32, which is what it does, and adjust the
callers to pass the value from the map lookup rather than making
parseUInt32 do the lookup itself.

Signed-off-by: Dave Cheney <dave@cheney.net>