internal/dag: improve the StatusWriter API
Previouly each ObjectStatusWriter had the ability to commit its status
directly via the Commit method. Commit is a final state, it can only be
set once per dag run. This API was confusing, the intention was the
creater of the ObjectStatusWriter, via sw.WithObject(), should be the
one to call Commit however this was not expressed well in the API and
obviously the failure to call Commit would discard any status set during
dag build.

The new API returns a pair of values, the ObjectStatusWriter and a
commit function. The caller should pass the ObjectStatusWriter to
functions interested in writing status, but keep the commit function for
itself. The commit function should be either called via a defer, or
directly if status's are being set in a loop (as defers will not fire
until the end of the function).

Signed-off-by: Dave Cheney <dave@cheney.net>