internal/dag: do not set VHost on non root IngressRoutes
This PR removes the Vhost setting on non root ingressroute records.
The included test shows that the Vhost status on a delegated
ingressroute is unstable, it depends on the order in which the
ingressroute's that delegate to it are processed and ultimaltely is
incorrect in both cases -- status only lets us set the Vhost field once,
yet more than one root ingressroute with different fqdns delegates to
ir31.

Signed-off-by: Dave Cheney <dave@cheney.net>