internal/envoy: move x-request-start header to RouteConfiguration
This PR fell out of debugging #1514.

A year ago #730 added the x-request-start header to each incoming
request. At the time this was done by adding the hedaer on each
route. This is verbose to see in the configuration and can be done
once at the enclosing RouteConfiguration.

This change means empty RouteConfigurations will have a
RequestHeadersToAdd field, but this is of little consiquence as it
applies to no routes as the list of virtualhosts is empty.

Another side effect of this change is the x-request-start header will be
added to redirects, however as this is a _request_ header, it is never
presented downstream during 301 processing so again it should have
little impact.

Signed-off-by: Dave Cheney <dave@cheney.net>