internal/dag: restore HTTPProxy processing
Restore most of the HTTPProxy processing. Restored builder_test.go,
but not status_test.go.

It turns out that collecting routes and their conditions and processing
those routes are tightly coupled because to resolve a service that a
route points to, you need the context of the specific object that holds
the route because includes allow routes to span namespaces.

Signed-off-by: Dave Cheney <dave@cheney.net>