internal/e2e: port over assertEquals fix from featuretests
The featuretests compare two proto's in a way that isn't impossible to
debug. Even though e2e is going away, port this over to make it easier
to debug bringing the HTTPProxy tests back online.

Nick: yes, this would be the third time we've written this code.
Probably time for a helper package.

Signed-off-by: Dave Cheney <dave@cheney.net>