internal/dag: validate TLS secrets during cache insertion
Fixes #1065

Extend the secret validation to inspect the contents of the secret
object to try to avoid sending malformed certificate data to envoy.

For this PR log the failure at the point of insertion. This should keep
the log volume low, unless some automated process is hammering the
secret with noop edits.

The result of rejecting the secret at cache insertion time is
SecureVirtualHosts or HTTPProxies that rely on certificate validation
will no longer be able to find their previously broken secrets and will
be excluded during the build. This is unlikely to be noticable by end
users as an invalid (not unsigned or expired, but garbage) certificate
would previously crash Envoy.

Signed-off-by: Dave Cheney <dave@cheney.net>