api: move HTTPProxy health check to route level
Updates #1119

This PR moves the HealthCheck stanza from the individual service level
to the route. This is an insurance policy if we have to move from one
cluster per service to using one service per route and multiple
clusterloadassignments. If that happens, we will not be able to support
per service configuration.

Specifically:

- Fork httpproxy/v1/healthCheck into ingressroute/v1beta/healthCheck
retaining the original feature set.
- Rename httpproxy/v1/healthCheck to healthCheckPolicy, making it
consistent with timeoutPolicy and retryPolicy.
- The underlying type is renamed to HTTPHeathCheckPolicy as it is
attached to Route.
- Healthchecks on httproxy.tcpproxy are removed. We can add them back in
rc.1 or rc.2.

Signed-off-by: Dave Cheney <dave@cheney.net>