api: remove HTTPProxy.Route.PrefixRewrite
Fixes #1667
Updates #899

This PR removes route.prefixRewrite from HTTPProxy in rc.1. The plan is
to replace it with something better in rc.2. We are removing this now in
rc.1 to avoid blocking moving HTTPProxy to v1 status in rc.1.

Please follow #899 for the replacement design which is scheduled for
rc.2

Signed-off-by: Dave Cheney <dave@cheney.net>