all: migrate to networking.k8s.io/v1beta1.Ingress
Fixes #1633

This PR changes Contour to use the networking/v1beta1.Ingress version of
the Ingress object. Contour uses this object internally and relies on
the API server to convert the older extensions/v1beta1 form to the current
networking/v1beta1 form.  Where this translation is not possible because
the cluster is too old, Contour performs this translation internally.

The catch is when relying on this translation mechanism, registering for
*both* extensions/v1beta1 and networking/v1beta1 causes objects that are
natively in the former format to appear twice. Thus, contour can only
subscribe to one of extensions/v1beta1 or networking/v1beta1. Thus
this PR provides a new flag, --use-extensions-v1beta1-ingress to
subscribe to the older form in pre 1.14 clusters.

Signed-off-by: Dave Cheney <dave@cheney.net>