internal/contour: order the route header match list
Ensure that the HeaderMatch slice on each Route is ordered. This
lets us endure that Route messages which have the same prefix and
number of header matches sort stably and will not cause unnecessary
Envoy API updates.

The previous code ordered the routes by first by the length of the
header match slice and then by the length of the prefix or regex
match. This change reverses that, but there is no net change in how
routes will be matched.

This fixes #1579.

Signed-off-by: James Peach <jpeach@vmware.com>