site: point to latest github release tag where possible
Fixes #1619

For 1.0 the plan of record is to keep the examples on GitHub. This is
due to the difficulty in doing directory index style static resources
from projectcontour.io.

To support this weave the name of the current github release through the
website.

Note that at the present, in the rc.2 timeframe, this will break a lot
of links so this PR is targeted for 1.0 (or shortly after) which will
bring the guides, resources, and release documentation in line with the
latest released tag on GitHub.

Signed-off-by: Dave Cheney <dave@cheney.net>