site: automatically link to the latest release.
Improve the "download now" CTA link to point to the latest release,
rather than the /releases landing page.

Because Jekyll cannot use templating variables in front matter move the
second call to action description in line in the html. For consistency
move the first call to action in line and use the {% link helper to
ensure it always points to a valid page.

Signed-off-by: Dave Cheney <dave@cheney.net>