site: use {% link in top nav
Use {% link consistently throughout the top navigation. This will
prevent the site from building if we remove/rename one of those pages
and has the added benefit that {% link generates the correct link rather
than bouncing through a 301 redirect to append a / to the url.

Lastly, remove the unused class tags on the li element.

Signed-off-by: Dave Cheney <dave@cheney.net>