internal/dag: accept non-PEM data in PEM files
It is valid for files to intermingle PEM and non-PEM data. The Go
`pem.Decode` API supports this, but doesn't emit any errors that
would let us distinguish between broken PEM and missing PEM. So,
verify that there is a remaining PEM header before we try to decode
it. That way, we know that any decoding failure is a actual error.

This fixes #1809.

Signed-off-by: James Peach <jpeach@vmware.com>