internal/dag: check annotation applicability
When Contour makes use of annotations, it just checks the object
kind where the annotation ought to be. If a user has accidentally
placed the annotation on the wrong kind of object, there is no
indication that something is wrong.

This change adds some initial validation that annotations are applied
to the correct kind of object and defends against unknown annotations
in the Contour annotation namespace. If we detect a problem, we
just log a warning, but eventually we should place an indicator in
the object status so that the user has better visibility.

This fixes #1825.

Signed-off-by: James Peach <jpeach@vmware.com>